Add a base class for common methods
