+++ app/lib/kgcl/renderers/node_rename_content.rb
@@ -9,52 +9,25 @@ module KGCL
     #
     # @see https://github.com/INCATools/kgcl KGCL documentation
     #
-    class NodeRenameContent
-      attr_reader :params
-
-      def initialize(params)
-        @params = params
-      end
-
-      def render
-        tr = KGCL::TemplateRenderer.new(
-          title_template: 'node_rename_title.erb',
-          body_template: 'node_rename_body.erb',
-          bind_klass: self
-        )
-        tr.render
-      end
-
+    class NodeRenameContent < IssueContent
       def comment
         @params[:node_rename][:comment]
-      end
-
-      def concept_label
-        @params[:concept_label]
-      end
-
-      def curie
-        @params[:curie]
       end
 
       def get_binding
         binding
       end
 
-      def github_id
-        @params[:github_id]
-      end
-
-      def orcid_id
-        @params[:orcid_id]
-      end
-
       def new_concept_label
         @params[:node_rename][:new_preferred_name]
       end
 
-      def username
-        @params[:username]
+      def title_template
+        'node_rename_title.erb'
+      end
+
+      def body_template
+        'node_rename_body.erb'
       end
     end
   end
